Crate multihash_codetable

A batteries-included code table of multihashes.

Whilst the multihash crate itself only defines the data structure, this crate defines a codetable via multihash_derive for several hashers. Although convenient, depending on this crate will increase the dependency footprint of your project.

It is only recommended as a getting-started-quickly solution. For production, you should either define your own codetable or rely only on the multihash crate itself.

Structs

Blake2bHasher

Multihash hasher.

Keccak224

Multihash hasher.

Keccak256

Multihash hasher.

Keccak384

Multihash hasher.

Keccak512

Multihash hasher.

Sha2_256

Multihash hasher.

Sha2_512

Multihash hasher.

Sha3_224

Multihash hasher.

Sha3_256

Multihash hasher.

Sha3_384

Multihash hasher.

Sha3_512

Multihash hasher.

Enums

Code

Default (cryptographically secure) Multihash implementation.

Traits

MultihashDigest

Trait that implements hashing.

Type Aliases

Blake2b256

256 bit blake2b hasher.

Blake2b512

512 bit blake2b hasher.

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.

Derive Macros

MultihashDigest

Custom derive for the MultihashDigest trait.